python-services/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openpyxl>=3.1.0
python-multipart>=0.0.12
faker>=22.0.0
pydantic>=2.0.0
PyYAML>=6.0
httpx>=0.25.0
websockets>=12.0
requests>=2.31.0


